fix(config): narrow BreakawayGen MaxGenerations to uint8
Mirrors the Rust u8 it round-trips to. Before this change a Go-authored

config could set MaxGenerations: 300 and silently truncate or fail when

the config reached the engine.

Mirrors the bf17e39 precedent for GenerationCommissionConfig.MaxGenerations

and adds the same boundary-table regression test pattern from db07634.

+++ internal/config/types.go
@@ -326,8 +326,12 @@ type FixedOverrideConfig struct {
 }
 
 // BreakawayGenerationConfig holds generation override configuration for breakaway groups.
+//
+// MaxGenerations uses uint8 to mirror the Rust u8 field exactly. This makes
+// negative or out-of-range values fail at Go unmarshal time instead of silently
+// truncating or erroring downstream when the config round-trips to the engine.
 type BreakawayGenerationConfig struct {
-	MaxGenerations  int                `yaml:"max_generations" json:"max_generations"`
+	MaxGenerations  uint8              `yaml:"max_generations" json:"max_generations"`
 	GenerationRates map[string]float64 `yaml:"generation_rates" json:"generation_rates"`
 	BoundaryRank    string             `yaml:"boundary_rank" json:"boundary_rank"`
 }
